feat(INV-38): M2-T2 候选筛选器 — Screener + configurable rules

+++ src/db/models.py
@@ -1,6 +1,6 @@
 from datetime import datetime
 
-from sqlalchemy import BigInteger, Date, DateTime, Index, Integer, Numeric, String, UniqueConstraint, func
+from sqlalchemy import Date, DateTime, Index, Integer, Numeric, String, UniqueConstraint, func
 from sqlalchemy.orm import Mapped, mapped_column
 
 from src.db.base import Base
@@ -82,7 +82,7 @@ class DerivedFactor(Base):
         Index('ix_derived_factors_factor_name_asof_date', 'factor_name', 'asof_date'),
     )
 
-    id: Mapped[int] = mapped_column(BigInteger, primary_key=True, autoincrement=True)
+    id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
     symbol: Mapped[str] = mapped_column(String(32), nullable=False)
     market: Mapped[str] = mapped_column(String(16), nullable=False)
     asof_date: Mapped[datetime] = mapped_column(Date, nullable=False)
